redid forecast model to ignore error implementing custom cell added a city label and the zip code helper added more tests added place mark coordination.

+++ WeatherApp/WeatherApp/Models/Forecast.swift
@@ -119,7 +119,7 @@ struct DailyDatum: Codable {
     let cloudCover: Double
     let uvIndex: Int
     let uvIndexTime: Int
-    let visibility: Int
+    let visibility: Double
     let ozone: Double
     let temperatureMin: Double
     let temperatureMinTime: Int
